+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- added the property `partitionsConsumedConcurrently` to the command executor
+
 ## [1.0.4] 2022-10-26
 
 ### Changed